master: cambios finales implementación

+++ Solstice/Models/Contact.cs
@@ -7,7 +7,7 @@
 
     public class Contact
     {
-        public double Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Company { get; set; }
         public string Image { get; set; }
